+++ examples/oven-app/silabs/BUILD.gn
@@ -106,10 +106,8 @@ silabs_executable("oven_app") {
   sources = [
     "${example_oven_dir}/oven-app-common/src/CookEndpoints.cpp",
     "${example_oven_dir}/oven-app-common/src/OvenEndpoint.cpp",
-    "${examples_common_plat_dir}/customer/CustomerAppManager.cpp",
     "${examples_common_plat_dir}/customer/CustomerAppTask.cpp",
     "src/AppTask.cpp",
-    "src/OvenManager.cpp",
   ]
   if (!disable_lcd) {
     sources += [ "src/OvenUI.cpp" ]
