--- examples/oven-app/silabs/include/AppManagerImpl.h
@@ -1,28 +1,0 @@
-/*
- *
- *    Copyright (c) 2025 Project CHIP Authors
- *    All rights reserved.
- *
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License.
- *    You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS,
- *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *    See the License for the specific language governing permissions and
- *    limitations under the License.
- */
-
-#pragma once
-
-/**
- * @file
- * @brief Per-app alias header consumed by shared CustomerAppManager.
- *
- * The shared CustomerAppManager includes "AppManagerImpl.h"
- */
-
-#include "OvenManagerImpl.h"
