--- examples/platform/silabs/customer/CustomerAppManager.cpp
@@ -1,32 +1,0 @@
-/*
- *
- *    Copyright (c) 2025 Project CHIP Authors
- *    All rights reserved.
- *
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License.
- *    You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS,
- *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *    See the License for the specific language governing permissions and
- *    limitations under the License.
- */
-
-/**
- * @file
- * @brief Customer-facing <AppName>Manager definition site, shared across apps.
- *
- * Add `*Impl()` overrides here to customize individual <AppName>Manager behaviors.
- * Any `*Impl()` you do not override keeps the default <AppName>Manager behavior.
- *
- * See the app README ("Override API Reference") for the full list of
- * overridable methods.
- */
-
-#include "CustomerAppManager.h"
-
-CustomerAppManager CustomerAppManager::sInstance;
