--- examples/platform/silabs/customer/CustomerAppManager.h
@@ -1,41 +1,0 @@
-/*
- *
- *    Copyright (c) 2025 Project CHIP Authors
- *    All rights reserved.
- *
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License.
- *    You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS,
- *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *    See the License for the specific language governing permissions and
- *    limitations under the License.
- */
-
-#pragma once
-
-#include "AppManagerImpl.h"
-
-/**
- * @brief Minimal <AppName>ManagerImpl-derived leaf, shared across apps.
- *
- * Any *Impl() method you do not override keeps the default <AppName>Manager behavior.
- * Override *Impl() methods here to customize individual behaviors.
- *
- * See the app README ("Override API Reference") for the full list of
- * overridable methods.
- */
-class CustomerAppManager : public AppManagerImpl<CustomerAppManager>
-{
-public:
-    static CustomerAppManager & GetInstance() { return sInstance; }
-
-private:
-    friend AppManagerImpl<CustomerAppManager>;
-
-    static CustomerAppManager sInstance;
-};
